requirements.txt
streamlit
ibm-watson
transformers
pandas
numpy
matplotlib
python-dotenv
requests